Web_development/requirements.txt
Flask==2.3.3
mysql-connector-python==8.0.33
werkzeug==2.3.3
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==9.6.1
